+++ src/main/java/dev/lambdaurora/lambdynlights/LambDynLightsCompat.java
@@ -17,7 +17,7 @@ import net.fabricmc.loader.api.VersionParsingException;
  * Represents a utility class for compatibility.
  *
  * @author LambdAurora
- * @version 3.0.0
+ * @version 3.1.0
  * @since 1.0.0
  */
 public final class LambDynLightsCompat {
@@ -28,6 +28,10 @@ public final class LambDynLightsCompat {
 	 */
 	public static boolean isCanvasInstalled() {
 		return FabricLoader.getInstance().isModLoaded("canvas");
+	}
+
+	public static boolean isSodiumInstalled() {
+		return FabricLoader.getInstance().isModLoaded("sodium");
 	}
 
 	public static boolean isSodium05XInstalled() {