+++ src/main/java/dev/lambdaurora/lambdynlights/mixin/lightsource/FallingBlockEntityMixin.java
@@ -1,0 +1,49 @@
+/*
+ * Copyright © 2024 LambdAurora <email@lambdaurora.dev>
+ *
+ * This file is part of LambDynamicLights.
+ *
+ * Licensed under the Lambda License. For more information,
+ * see the LICENSE file.
+ */
+
+package dev.lambdaurora.lambdynlights.mixin.lightsource;
+
+import dev.lambdaurora.lambdynlights.DynamicLightSource;
+import dev.lambdaurora.lambdynlights.LambDynLights;
+import dev.lambdaurora.lambdynlights.api.DynamicLightHandlers;
+import net.minecraft.world.entity.Entity;
+import net.minecraft.world.entity.EntityType;
+import net.minecraft.world.entity.decoration.BlockAttachedEntity;
+import net.minecraft.world.entity.item.FallingBlockEntity;
+import net.minecraft.world.level.Level;
+import net.minecraft.world.level.block.state.BlockState;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Shadow;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+
+@Mixin(FallingBlockEntity.class)
+public abstract class FallingBlockEntityMixin extends EntityMixin implements DynamicLightSource {
+	@Shadow
+	private BlockState blockState;
+
+	@Inject(method="tick", at = @At("RETURN"))
+	public void lambdynlights$onTick(CallbackInfo ci) {
+		// We do not want to update the entity on the server.
+		if (this.level().isClientSide()) {
+			if (this.isRemoved()) {
+				this.setDynamicLightEnabled(false);
+			} else {
+				this.dynamicLightTick();
+
+				this.lambdynlights$luminance = Math.max(this.lambdynlights$luminance, this.blockState.getLightEmission());
+
+				if (!LambDynLights.get().config.getEntitiesLightSource().get() || !DynamicLightHandlers.canLightUp((Entity) (Object) this))
+					this.lambdynlights$luminance = 0;
+				LambDynLights.updateTracking(this);
+			}
+		}
+	}
+}